lint fixes and test import refactoring -> clean tests
