+++ src/components/Spinner.js
@@ -24,4 +24,4 @@ export default class Spinner extends Component {
             </div>
         )
     }
-};
+}
