+++ src/components/diagnostic/DiagnosticPane.js
@@ -17,12 +17,6 @@ class DiagnosticPane extends Component {
         return (
             <div className="DiagnosticPane" key={this.state.key}>
                 <Grid divided='vertically'>
-                    {/* <Grid.Row columns={1}>
-                        <Grid.Column>
-                            <Overview key={this.state.key} node={this.props.node}/>
-                        </Grid.Column>
-                    </Grid.Row> */}
-
                     <Grid.Row columns={1}>
                         <Grid.Column>
                             <h3>Run Diagnostics <Explainer knowledgebase='Diagnostics'/></h3>
