+++ src/components/timeseries/CypherTimeseries.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import CypherTimeseries from './CypherTimeseries';
 import uuid from 'uuid';
-import fakes from '../testutils/fakes';
+import fakes from '../../testutils/fakes';
 
 describe('Cypher Timeseries', function() {
     const key = uuid.v4();
